pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "insurance-pdf-extractor"
version = "0.1.0"
description = "Tool for extracting data from insurance PDFs"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Project Author", email = "author@example.com"},
]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]

dependencies = [
    # List your dependencies here, for example:
    # "fastapi-mcp>=0.1.0",
    "fastapi-mcp>=0.3.4",
    "google-generativeai>=0.8.5",
    "pypdf>=5.5.0",
    "pytest>=8.3.5",
    "pyyaml",
    "slowapi>=0.1.9",
]

[project.urls]
"Homepage" = "https://github.com/yourusername/insurance-pdf-extractor"
"Bug Tracker" = "https://github.com/yourusername/insurance-pdf-extractor/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/insurance_pdf_extractor"]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.0.0",
]

[tool.black]
line-length = 88

[tool.isort]
profile = "black"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true

[tool.pytest.ini_options]
addopts = "-v --tb=short"
testpaths = ["tests"]
asyncio_mode = "auto"
markers = [
    "asyncio: marks tests as async (deselect with '-m \"not asyncio\"')",
]

